new version, added fees changes, nfloats, validation and etc
Former-commit-id: 9bed96a96b8f01f1e7a0819b9e079ad0d569ba34 [formerly 001435701f0fdbab852fad9b600a1919665d75c3]
Former-commit-id: fd376c3a8950f7bb6b8756fe47471ce98382f5fc

+++ transactions/transactions.js
@@ -38,8 +38,8 @@ transaction.prototype.getBytes = function () {
         bb.writeByte(recepient[i] || 0);
     }
 
-    bb.writeFloat64(this.amount);
-    bb.writeFloat64(this.fee);
+    bb.writeLong(this.amount);
+    bb.writeLong(this.fee);
 
     if (this.referencedTransaction) {
         var referncedTransactionBuffer = bignum(this.referencedTransaction).toBuffer();
@@ -116,8 +116,8 @@ transaction.prototype.fromBytes = function (buffer) {
 
     var recepient = bb.readLong();
     t.recipientId = recepient + "C";
-    t.amount = bb.readInt();
-    t.fee = bb.readInt();
+    t.amount = bb.readUint64();
+    t.fee = bb.readUint64();
     t.referencedTransaction = bb.readLong();
 
     var signature = new Buffer(64);
